whitespace

+++ stack1.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <stdint.h>
+
 #define STACK_SIZE	(5)
 
 typedef struct {
-	char* string[STACK_SIZE];
-	unsigned int counter;
+    char* string[STACK_SIZE];
+    uint16_t counter;
 }stack_t;
 
 extern stack_t stack;
